feat: orthodox slotfill moved to slotfill folder
