+++ deeppavlov/__init__.py
@@ -42,7 +42,6 @@ import deeppavlov.models.embedders.fasttext_embedder
 import deeppavlov.models.embedders.dict_embedder
 import deeppavlov.models.embedders.glove_embedder
 import deeppavlov.models.embedders.bow_embedder
-import deeppavlov.models.ner.slotfill
 import deeppavlov.models.ner.ner_ontonotes
 import deeppavlov.models.spellers.error_model.error_model
 import deeppavlov.models.trackers.hcn_at
@@ -75,6 +74,7 @@ import deeppavlov.models.preprocessors.field_getter
 import deeppavlov.models.preprocessors.sanitizer
 import deeppavlov.models.preprocessors.lazy_tokenizer
 import deeppavlov.models.slotfill.slotfill_raw
+import deeppavlov.models.slotfill.slotfill
 
 
 import deeppavlov.metrics.accuracy
